% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\docType{package}
\name{ECAMECO}
\alias{ECAMECO}
\alias{ECAMECO-package}
\title{A package for downloading the European Commission's AMECO Database.}
\description{
The workhorse function of the package is \code{getECAMECO()}. This function returns a
list containing the AMECO database. The list contains a list for each topic in the database.
The code behind this package also serves as a backend to an EViews add-in created by the author.
}
\section{ECAMECO functions}{

\code{getECAMECO()}
}

