% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getECAMECO}
\alias{getECAMECO}
\title{Get European Commission AMECO Database}
\usage{
getECAMECO(url = ECAMECO.env$ameco0_url)
}
\arguments{
\item{url}{The URL pointing to the .zip file}
}
\description{
Downloads the AMECO Database into R
}
\details{
The url parameter is set by default to the latest version of the AMECO database. This function can also downloand the vintage files. The vintage URLs can be found using the \code{getVintagesURLs()} function.
}
\author{
Graeme Walsh
}
