% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getHomepage}
\alias{getHomepage}
\title{Get the AMECO database homepage (.html file)}
\usage{
getHomepage()
}
\description{
Reads the AMECO database html page into R
}
\details{
The homepage is (https://ec.europa.eu/info/business-economy-euro/indicators-statistics/economic-databases/macro-economic-database-ameco/download-annual-data-set-macro-economic-database-ameco_en)
}
\author{
Graeme Walsh
}
