% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getLastUpdate}
\alias{getLastUpdate}
\title{Get the last update of the database}
\usage{
getLastUpdate()
}
\description{
Extracts the last update value from the AMECO database html page
}
\details{
The homepage is (https://ec.europa.eu/info/business-economy-euro/indicators-statistics/economic-databases/macro-economic-database-ameco/download-annual-data-set-macro-economic-database-ameco_en)
}
\author{
Graeme Walsh
}
