% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getURL}
\alias{getURL}
\title{Get the URL for the latest .zip file}
\usage{
getURL()
}
\description{
Extracts the latest .zip file URL from the AMECO database html page
}
\details{
The homepage is (https://ec.europa.eu/info/business-economy-euro/indicators-statistics/economic-databases/macro-economic-database-ameco/download-annual-data-set-macro-economic-database-ameco_en)
}
\author{
Graeme Walsh
}
