% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getVintagespage}
\alias{getVintagespage}
\title{Get the AMECO database vintages page (.html file)}
\usage{
getVintagespage()
}
\description{
Reads the AMECO database vintages html page into R
}
\details{
The vintages page is (http://ec.europa.eu/economy_finance/db_indicators/ameco/archive_en.htm)
}
\author{
Graeme Walsh
}
