% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getVitageURLs}
\alias{getVitageURLs}
\title{Get the URL for the vintage .zip files}
\usage{
getVitageURLs()
}
\description{
Extracts the .zip file URLs from the AMECO database vintage page
}
\details{
The homepage is (http://ec.europa.eu/economy_finance/db_indicators/ameco/archive_en.htm)
}
\author{
Graeme Walsh
}
