% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECAMECO.R
\name{getZipURLs}
\alias{getZipURLs}
\title{Get the zipped files URLs}
\usage{
getZipURLs()
}
\description{
Extracts the zipped files URLs from the AMECO database html page
}
\details{
The homepage is (https://ec.europa.eu/info/business-economy-euro/indicators-statistics/economic-databases/macro-economic-database-ameco/download-annual-data-set-macro-economic-database-ameco_en)
}
\author{
Graeme Walsh
}
